+++ tools/cosmovisor/cmd/cosmovisor/add_upgrade.go
@@ -9,7 +9,6 @@ import (
 	"github.com/spf13/cobra"
 
 	"cosmossdk.io/tools/cosmovisor"
-	upgradetypes "github.com/cosmos/cosmos-sdk/x/upgrade/types"
 )
 
 func NewAddUpgradeCmd() *cobra.Command {
@@ -64,7 +63,7 @@ func addUpgrade(cfg *cosmovisor.Config, force bool, upgradeHeight int64, upgrade
 	logger.Info(fmt.Sprintf("Upgrade binary located at %s", cfg.UpgradeBin(upgradeName)))
 
 	if upgradeHeight > 0 {
-		plan := upgradetypes.Plan{
+		plan := cosmovisor.ManualUpgradePlan{
 			Name:   upgradeName,
 			Height: upgradeHeight,
 		}
